Fix build bug on GCC<4.6: duplicate typedefs
On versions of GCC which don't support the c11 standard, duplicate typedefs,
even if they are exactly the same, are not permitted.

On GCC 4.5.2, we get a build error when led.h and stdint.h are included in the
same file:

    In file included from ../system/platforms/arm-qemu/platforminit.c:14:0:
    ../include/led.h:50:23: error: redefinition of typedef 'uint8_t'
    ../system/platforms/arm-qemu/stdint.h:15:23: note: previous declaration of 'uint8_t' was here
    ../include/led.h:51:22: error: redefinition of typedef 'uint32_t'
    ../system/platforms/arm-qemu/stdint.h:21:22: note: previous declaration of 'uint32_t' was here

Newer versions of GCC won't generate a warning or error for this, unless passed
'-pedantic' or '-pedantic -Werror'. There are too many '-pedantic' warnings
currently to make adding '-pedantic' practical in this commit.