Update lexer/parser for breaking changes in flex
New versions of flex (>2.5.35) change the type of yyleng from an 'int'
to a 'yy_size_t'. The new type is quite reasonable, since lengths will
never be negative, but it broke a lot of things in the config parser.

I've modified the build so that flex produces a header file for the
parser to use. Unfortunately, I couldn't stop there. Since the parser
include's the entire scanner generated by flex (yy.lex.c) and the
generated scanner declares types differently from the generated header
file I had to move away from having the scanner contained in the parser.

Separating them was easy enough and the change to the build system was
pretty simple. However, as written, the parser depened on a static
function in the generated scanner which -- after the separation -- is no
longer visible to the parser. I was able to move that into the lexer,
however, so everything works again.

Eventually I'd like to move to a solution where I don't need to keep the
exported char* verbatim_out shared between the lexer and the parser (I
dislike using globals like that), but it's a minor issue that I can work
out later.