Sloppy workaround for the VIC problem in QEMU.
In some versions of QEMU, there's an off-by-one error in the vectored
interrupt controller (specifically, the PL190). This was fixed in
September 2012, but many systems have not updated to that fix yet.

The workaround is to figure out the source of interrupt and jump to the
correct handler manually. This is a lot more work than just reading from
VIC_VEC_ADDR and jumping, but should work in systems which do and do not
have VICs with the off-by-one error.

Bugs:

* The manual fallback only happens if the interrupt handler is null, so
  weird stuff will happen if we have two IRQs next to each-other
  enabled (because there will still be an off-by-one error, but we won't
  detect it). We should either auto-detect the bug or always do the
  workaround.

* This change doesn't handle interrupt priorization properly. In fact,
  it services interrupts in the reverse order that it's supposed to! It
  works fine for single interrupts.